# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OppAI Stream API - Scraper API for oppai.stream anime content
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface StreamQuality {
  /** Quality label (e.g. "1080p", "720p", "480p", "360p") */
  quality: string;
  /** Direct stream URL */
  url: string;
  /** Stream type (e.g. "m3u8", "mp4") */
  type: string;
}

export type EpisodeStreamsSubtitlesItem = {
  language: string;
  url: string;
};

export interface EpisodeStreams {
  animeSlug: string;
  animeName: string;
  episodeNumber: number;
  episodeTitle?: string;
  thumbnail?: string;
  streams: StreamQuality[];
  subtitles?: EpisodeStreamsSubtitlesItem[];
}

export interface EpisodeSummary {
  number: number;
  title?: string;
  thumbnail?: string;
  url: string;
}

export interface AnimeInfo {
  slug: string;
  title: string;
  titleJapanese?: string;
  coverImage?: string;
  bannerImage?: string;
  synopsis?: string;
  genres?: string[];
  /** Airing status (e.g. "Ongoing", "Completed") */
  status?: string;
  /** Type (e.g. "TV", "Movie", "OVA") */
  type?: string;
  studio?: string;
  releaseYear?: string;
  season?: string;
  rating?: string;
  /** Number of available episodes */
  episodeCount: number;
  episodes?: EpisodeSummary[];
}

export interface SearchResult {
  slug: string;
  title: string;
  coverImage?: string;
  genres?: string[];
  status?: string;
  type?: string;
  episodeCount?: number;
  url: string;
}

export interface SearchResults {
  query: string;
  page: number;
  results: SearchResult[];
  hasNextPage: boolean;
}

export interface EpisodesList {
  animeSlug: string;
  animeName: string;
  episodeCount: number;
  episodes: EpisodeSummary[];
}

export interface LatestEpisodeEntry {
  animeSlug: string;
  animeName: string;
  episodeNumber: number;
  episodeTitle?: string;
  thumbnail?: string;
  coverImage?: string;
  releasedAt?: string;
  url: string;
}

export interface LatestEpisodes {
  page: number;
  episodes: LatestEpisodeEntry[];
  hasNextPage: boolean;
}

export type SearchAnimeParams = {
  /**
   * Search query
   */
  q: string;
  /**
   * Page number
   * @minimum 1
   */
  page?: number;
};

export type GetLatestEpisodesParams = {
  /**
   * Page number
   * @minimum 1
   */
  page?: number;
};