# lib/api-zod/src/generated/types/animeInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OppAI Stream API - Scraper API for oppai.stream anime content
 * OpenAPI spec version: 1.0.0
 */
import type { EpisodeSummary } from "./episodeSummary";

export interface AnimeInfo {
  slug: string;
  title: string;
  titleJapanese?: string;
  coverImage?: string;
  bannerImage?: string;
  synopsis?: string;
  genres?: string[];
  /** Airing status (e.g. "Ongoing", "Completed") */
  status?: string;
  /** Type (e.g. "TV", "Movie", "OVA") */
  type?: string;
  studio?: string;
  releaseYear?: string;
  season?: string;
  rating?: string;
  /** Number of available episodes */
  episodeCount: number;
  episodes?: EpisodeSummary[];
}