# lib/api-zod/src/generated/types/episodeStreams.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OppAI Stream API - Scraper API for oppai.stream anime content
 * OpenAPI spec version: 1.0.0
 */
import type { EpisodeStreamsSubtitlesItem } from "./episodeStreamsSubtitlesItem";
import type { StreamQuality } from "./streamQuality";

export interface EpisodeStreams {
  animeSlug: string;
  animeName: string;
  episodeNumber: number;
  episodeTitle?: string;
  thumbnail?: string;
  streams: StreamQuality[];
  subtitles?: EpisodeStreamsSubtitlesItem[];
}