# lib/api-zod/src/generated/types/episodeSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OppAI Stream API - Scraper API for oppai.stream anime content
 * OpenAPI spec version: 1.0.0
 */

export interface EpisodeSummary {
  number: number;
  title?: string;
  thumbnail?: string;
  url: string;
}