# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OppAI Stream API - Scraper API for oppai.stream anime content
 * OpenAPI spec version: 1.0.0
 */

export interface HealthStatus {
  status: string;
}