# lib/api-zod/src/generated/types/latestEpisodeEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OppAI Stream API - Scraper API for oppai.stream anime content
 * OpenAPI spec version: 1.0.0
 */

export interface LatestEpisodeEntry {
  animeSlug: string;
  animeName: string;
  episodeNumber: number;
  episodeTitle?: string;
  thumbnail?: string;
  coverImage?: string;
  releasedAt?: string;
  url: string;
}