# lib/api-zod/src/generated/types/latestEpisodes.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OppAI Stream API - Scraper API for oppai.stream anime content
 * OpenAPI spec version: 1.0.0
 */
import type { LatestEpisodeEntry } from "./latestEpisodeEntry";

export interface LatestEpisodes {
  page: number;
  episodes: LatestEpisodeEntry[];
  hasNextPage: boolean;
}