# lib/api-zod/src/generated/types/searchResults.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OppAI Stream API - Scraper API for oppai.stream anime content
 * OpenAPI spec version: 1.0.0
 */
import type { SearchResult } from "./searchResult";

export interface SearchResults {
  query: string;
  page: number;
  results: SearchResult[];
  hasNextPage: boolean;
}