# lib/api-zod/src/generated/types/streamQuality.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OppAI Stream API - Scraper API for oppai.stream anime content
 * OpenAPI spec version: 1.0.0
 */

export interface StreamQuality {
  /** Quality label (e.g. "1080p", "720p", "480p", "360p") */
  quality: string;
  /** Direct stream URL */
  url: string;
  /** Stream type (e.g. "m3u8", "mp4") */
  type: string;
}